fix: update safe uint convertion to convert max int to uint first to validate

+++ commons/utils.go
@@ -158,10 +158,10 @@ func SafeInt64ToInt(val int64) int {
 
 // SafeUintToInt64 converts a uint to int64 safely by capping values at math.MaxInt64.
 func SafeUintToInt64(val uint) int64 {
-	if val > math.MaxInt64 {
-		return math.MaxInt64
+	if val > uint64(math.MaxInt64) {
+	    return math.MaxInt64
 	}
-
+ 	
 	return int64(val)
 }
